Move dtl and bstr headers into named directories
