+++ bstr/include/bstr/bstr.hpp
